requirements.txt
Django==1.9.5
djangorestframework==3.3.3
httpie==0.9.3
psycopg2==2.6.1
Pygments==2.1.3
requests==2.9.1
